Pull inventory diff function out of etsy file
